reduced all types to Float64 only removed module CodeTypes all methods of organisms code return Float64 fixed three issues with errors in generated code
#137

+++ src/net/RpcApi.jl
@@ -137,7 +137,7 @@ module RpcApi
     #
     # Organism's personal memory. Is used in any possible way.
     #
-    mem::Dict{Int16, Int16}
+    mem::Dict{Float64, Float64}
     #
     # Organism's position in a 2D world. Starts from (1,1)
     # ends with (worldWidth, worldHeight) configurations.
